Put class definition in other file
